# docs/swagger.yaml
basePath: /api/v1/cbesuperapp/cps_action
definitions:
...
  cbe-super-app-cps-action_internal_constants_dto_permission.CreatePermissionGroupRequest:
    properties:
      group_name:
        type: string
      permission_category_list:
        items:
          type: string
        type: array
      role:
        type: string
    type: object
  cbe-super-app-cps-action_internal_constants_dto_permission.PaginatedPermissionGroupResponse:
    properties:
      docs:
        items:
          $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_permission.PermissionCategoryResponse'
        type: array
      meta:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_types.PaginationMeta'
    type: object
  cbe-super-app-cps-action_internal_constants_dto_permission.PermissionCategoryResponse:
    properties:
      access:
        type: string
      categoryName:
        type: string
      createdAt:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      isDeleted:
        type: boolean
      permissions: {}
      updatedAt:
        type: string
    type: object
  cbe-super-app-cps-action_internal_constants_dto_permission.UpdatePermissionGroupRequest:
    properties:
      new_group_name:
        type: string
      old_group_name:
        type: string
      permission_category_list:
        items:
          type: string
        type: array
      role:
        type: string
    type: object
...
  cbe-super-app-cps-action_internal_constants_localization.StandardResponse:
    properties:
      data: {}
      message:
        type: string
      ok:
        type: boolean
      status:
        type: integer
      timestamp:
        type: string
    type: object
...
  cbe-super-app-cps-action_internal_constants_types.PaginationMeta:
    properties:
      has_next_page:
        type: boolean
      has_prev_page:
        type: boolean
      limit:
        type: integer
      next_page:
        type: integer
      page:
        type: integer
      paging_counter:
        type: integer
      prev_page:
        type: integer
      total_docs:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: contact@eaglelionsystems.com
    name: API Support
  description: API documentation for CPS Action service
  title: CPS Action API
  version: 1.0.0
paths:
...
  /permission-categories-with-permissions:
    get:
      description: Retrieves all permission categories with their permissions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_permission.PaginatedPermissionGroupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Get All Permission Categories With Permissions
      tags:
      - Permission
  /permission-groups:
    get:
      description: Retrieves a paginated list of permission groups
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_permission.PaginatedPermissionGroupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Get Permission Groups
      tags:
      - Permission
    post:
      consumes:
      - application/json
      description: Creates a new permission group
      parameters:
      - description: Create Permission Group DTO
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_permission.CreatePermissionGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Create Permission Group
      tags:
      - Permission
  /permission-groups/{group_name}:
    get:
      description: Retrieves a permission group by its name
      parameters:
      - description: Group Name
        in: path
        name: group_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_permission.PermissionCategoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Get Permission Group by Name
      tags:
      - Permission
    put:
      consumes:
      - application/json
      description: Updates an existing permission group
      parameters:
      - description: Group Name
        in: path
        name: group_name
        required: true
        type: string
      - description: Update Permission Group DTO
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_permission.UpdatePermissionGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Update Permission Group
      tags:
      - Permission
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"